// src/models/Supplier.ts
export interface Supplier {
    id: string
    code: string
    name: string
    address: string

    pic: string
    phone: string
    email: string

    bankName: string
    bankBranch: string
    bankAccountNumber: string

    taxIdentificationNumber: string
    taxableFirmName: string
    taxableFirmAddress: string
}